Update delete Doctor Popup

+++ build/web/receptionist.jsp
@@ -29,6 +29,23 @@
             #patientTable {
                 width: 100% !important;
             }
+            button.ajs-button.ajs-ok{
+                display: inline;
+            }
+            div.ajs-footer{
+                display: block;
+            }
+            div.ajs-content{
+                height: 100px;
+                font-size: 1.4em;
+                text-align: center;
+                color: #009efb;
+            }
+            div.ajs-dialog{
+                position: relative;
+                top: 70px;
+                width: 450px;
+            }
         </style>
     </head>
 
@@ -150,20 +167,21 @@
                             tr = table.getElementsByTagName("tr");
                             td = tr[rowindex + 1].getElementsByTagName("td")[3];
                             txtValue = td.textContent;
+                            
                             $.each(data, function (index, value) {
                                 if (value.phone === txtValue) {
                                     
                                     localStorage.setItem("dataRecep", JSON.stringify(value));
                                     $(document).on('click', '#delete', function () {
+                                        alertify.confirm("Are you sure?", function (asc) {
+                                        if (asc) {
                                          $.ajax({
                                             type: "DELETE",
                                             dataType: "json",
                                             contentType: "application/json; charset=utf-8",
                                             headers: {
                                             Authorization: 'Bearer ' + token},
-                                            beforeSend:function(){
-                                                return confirm("Are you sure?");
-                                             },
+                                            
                                             url: "http://14.161.47.36:8080/PHR_System-0.0.1-SNAPSHOT/accounts/account/" + value.accountId,
                                             complete: function (jqXHR) {
                                                 if (jqXHR.status === 200 || jqXHR.status === 201) {
@@ -171,11 +189,16 @@
                                                 }
                                             }
                                         });
+                                        
+                                        } else {
+                                            alertify.error("You've clicked cancel");
+                                        }
+                                        });
 
                                     });
                                 }
                             });
-
+                            
                         }
                         );
